+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Fix - (#6103) Fix opening forms dashboard if the user has invalid forms
   assigned (only show valid forms on dashboard)
+- Fix - (#6106) Form Session View - Fix link to form answers in form session
+  form answers table
 
 ## 7.3.0