+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Patches / Fixes
 
+- Fix - (#6101) Fix Dashboard Widget Display Issues
 - Fix - (#6103) Fix opening forms dashboard if the user has invalid forms
   assigned (only show valid forms on dashboard)
 - Fix - (#6106) Form Session View - Fix link to form answers in form session